Limit the visibility of the IO-id and adjust types.
This ID is not used outside of connect.c, so there is no point in having it in
the other structures. Also there is no warranty stated anywhere, that this ID is
in the range of int, so the check ID>0 can fail. Therefore the type is adjusted,
and the IO watch is always removed unconditionally (the watch create function
does not say that it can fail in any way).