Adjust the source to be compilable with OpenSUSE.
Due to a older gcc version on this platform the debug "log" function macro is
adjusted to use a older syntax of variadic macros.

The libc in OpenSuse is more strict about the exported symbols of the headers.
So the code is adjusted to enable those exports, and also include necessary
headers for string handling which are silently included in different headers on
other platforms.

Also there is a fix for a minor bug of seeding the wrong random number
generator.