Refactor the country and card lists.
The double storage of the country and card ownership is removed, which means
that it is impossible to have inconsistent states betewen the SPLAYER lists and
the information based on COUNTRY.numjug.

This gives the opportunity to refactor COUNTRY_FROM_TARJETA into a function
which does not depend on the storage layout of it's surrounding structure. The
drawback is, that the whole country array has to walked through where the list
code did work before. But since there are only about 50 countries, the
performance impact is not noticeable at the moment.

The COUNTRY and TARJETA	structures are now real C++ objects, which makes it
easier to initialize them.

As a side effect the boolean conditions of the FOW-part got reordered, so that
the short cut is now taken immediately.

Unfortunately I did not find a way to split this changeset into smaller parts.