Prepare the directory independence of the theme code.
Since there is some error when c++17 <filesystem> is included too late, this
commit does now add this include to the files which will be affected by the
transformation to the filesystem independend theme code. I choose to split this
commit from the following one in order to not have this clutter laying around
in the refactoring commit.